Move TLS settings from TcpTransport up into AbstractTcpTransport
